[fix] mount ./keys directory in jobs-runner
According to the readme located in the ./keys directory, you should be able to place a key pair there, and it will be used for git syncing. This did not work, because the jobs-runner was not mounting this folder, so it did not have access to the keys.

Added a volume mount to the container definition to fix this.